% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{member_importance}
\alias{member_importance}
\title{Calculate member importance in a graph}
\usage{
member_importance(gs, measure = "all")
}
\arguments{
\item{gs}{A list of graph objects}

\item{measure}{A character string of the centrality measure}
}
\value{
A data.frame. When measure is 'all'
\dontrun{
\item module The names in modules
\item names The names of the entries in modules
\item degree The degree centrality of each node, as in centr_degree
\item betweenness The betweennes centrality of each node, as in centr_betw
\item closeness The closeness centrality of each node, as in centr_clo
\item hub The hub score of each node, as in hub_score
}
}
\description{
Calculate several centrality measures on the graphs
}
