% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{member_similarity}
\alias{member_similarity}
\title{Calculate pair-wise similarity in an expresison matrix}
\usage{
member_similarity(dat, measure = "all")
}
\arguments{
\item{dat}{A data.frame of transposed expression matrix}

\item{measure}{A character string of the similarity measure}
}
\value{
A data.fame
}
\description{
Calculate pair-wise similarity in an expresison matrix
}
