% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{metadata_get}
\alias{metadata_get}
\title{Get metadata from ExpressionSet}
\usage{
metadata_get(eset, col_name, new_name, type = "phenotype")
}
\arguments{
\item{eset}{An ExpressionSet object}

\item{col_name}{A character vector of columns in the metadata}

\item{new_name}{A character string of the new names of the columns}

\item{type}{A character string (default 'phenotype')}
}
\value{
A data.frame always contains a column for each col_name with the
name in new_name. When type is 'phenotype', row.names are the samples,
and when type is 'features', row.names are the feature IDs/probes
}
\description{
Get the metadata associated with an ExpressionSet object. This could be
the phenotype or the features data. In either cases the functions uses
the standard bioconductor accessor with some formating
}
\examples{
library(Biobase)
data("sample.ExpressionSet")

metadata_get(sample.ExpressionSet,
             c('sex', 'type'),
             c('sex', 'type'))

}
