% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotation_get}
\alias{annotation_get}
\title{Get GO annotations}
\usage{
annotation_get(go_id, go_names, go_db, go_term, org_db, columns,
  remove_predicted = TRUE, ...)
}
\arguments{
\item{go_id}{A character vector of GO IDs}

\item{go_names}{A character vector of GO category names}

\item{go_db}{A GO.db object}

\item{go_term}{A GO.db object for GO terms}

\item{org_db}{An organism object}

\item{columns}{A charachter vector}

\item{remove_predicted}{A logical (default, TRUE) for removing predicted genes}

\item{...}{Other arguments passed to select}
}
\value{
A data.frame
}
\description{
Get GO annotations
}
