% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cna_run}
\alias{cna_run}
\title{Run WGCNA}
\usage{
cna_run(dat, power)
}
\arguments{
\item{dat}{A data.frame of transposed expression matrix}

\item{power}{An integer power of the adjacency matrix}
}
\value{
A list
}
\description{
Run WGCNA
}
\examples{
library(Biobase)
data(sample.ExpressionSet)
dat <- list(data = t(exprs(sample.ExpressionSet)))

cna_run(dat$data, power = 5)

}
