% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expression_reshape}
\alias{expression_reshape}
\title{Reshape data for WGCNA}
\usage{
expression_reshape(mat, log = FALSE, zero_neg = FALSE)
}
\arguments{
\item{mat}{A numeric expression matrix}

\item{log}{A logical (default, FALSE) to take the log matrix}

\item{zero_neg}{A logical (default, FALSE) to zero all negative values}
}
\value{
A list with one data.frame item, data contains the transposed matrix.
This is the genes in column and samples in rows
}
\description{
Reshap an expression matrix to the format required by WGCNA
}
\examples{
library(Biobase)
data(sample.ExpressionSet)

mat <- exprs(sample.ExpressionSet)

expression_reshape(mat)

}
