% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{member_importance}
\alias{member_importance}
\title{Calculate member importance in a graph}
\usage{
member_importance(gs, measure = "all")
}
\arguments{
\item{gs}{A list of graph objects}

\item{measure}{A character string of the centrality measure}
}
\value{
A data.frame
}
\description{
Calculate member importance in a graph
}
