% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{member_similarity}
\alias{member_similarity}
\title{Calculate pair-wise similarity in an expresison matrix}
\usage{
member_similarity(dat, measure = "all")
}
\arguments{
\item{dat}{A data.frame of transposed expression matrix}

\item{measure}{A character string of the similarity measure}
}
\value{
A data.fame. When measure is 'all'
\itemize{
\item x1 A gene ID
\item x2 A gene ID, as in edge list (x1 --> x2)
\item value The value of the measure
\item measure either Adjacency, TOM or Pearson
}
}
\description{
Calculate pair-wise similarity in an expresison matrix
}
