% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{metadata_get}
\alias{metadata_get}
\title{Get metadata from ExpressionSet}
\usage{
metadata_get(eset, col_name, new_name, type = "phenotype")
}
\arguments{
\item{eset}{An ExpressionSet object}

\item{col_name}{A character vector of columns in the metadata}

\item{new_name}{A character string of the new names of the columns}

\item{type}{A character string (default 'phenotype')}
}
\value{
A data.frame
}
\description{
Get metadata from ExpressionSet
}
\examples{
library(Biobase)
data("sample.ExpressionSet")

metadata_get(sample.ExpressionSet,
             c('sex', 'type'),
             c('sex', 'type'))

}
