% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{module_compare}
\alias{module_compare}
\title{Compare enrichment by modules}
\usage{
module_compare(index, level = 4, ...)
}
\arguments{
\item{index}{A list}

\item{level}{An integer}

\item{...}{Other arguments passed to compareCluster}
}
\value{
A data.frame
}
\description{
Compare enrichment by modules
}
